+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+<a name="2.2.0"></a>
+
+# [2.2.0](https://gitlab.com/atjdev/ezio-api/-/merge_requests/124) (2022-04-21)
+
+### Features
+
+- [**M014**] **warping**: add entity and resolver.
+- Add ifs-shop-material-alloc entity and resolver.
+- Add ifs-dop-head entity and resolver.
+- Add ifs-dop-supply-shop-ord entity and resolver.
+- Add query getStockByContractPartLocation at ifs-inv-part-in-stock resolver.
+- Add query getAttrValue at ifs-inv-part resolver.
+
 <a name="2.1.1"></a>
 
 # [2.1.1](https://gitlab.com/atjdev/ezio-api/-/merge_requests/122) (2022-04-20)